I've fixed the Vercel build error. I updated the build command and removed the `distDir` configuration.
