+++ .gitignore
@@ -1,0 +1,20 @@
+# Dependencies
+/node_modules
+
+# Vite
+.vite/
+dist/
+
+# Logs
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
+
+# Editor directories and files
+.vscode
+.idea
+*.suo
+*.ntvs*
+*.njsproj
+*.sln
+*.sw?
