+++ eslint.config.js
@@ -1,0 +1,8 @@
+export default [
+  {
+    rules: {
+      'no-unused-vars': 'warn',
+      'no-console': 'off',
+    },
+  },
+];
