+++ src/styles/testimonials.css
@@ -1,0 +1,73 @@
+/* Testimonials Section */
+.testimonials-section {
+  background-color: var(--light-color);
+  position: relative;
+}
+
+.testimonials-container {
+  display: grid;
+  grid-template-columns: repeat(auto-fit, minmax(320px, 1fr));
+  gap: 3rem;
+}
+
+.testimonial {
+  background: #2c2c2c;
+  border-radius: var(--border-radius);
+  padding: 2.5rem;
+  box-shadow: var(--box-shadow);
+  transition: var(--transition);
+  position: relative;
+}
+
+.testimonial:hover {
+  transform: translateY(-10px);
+  box-shadow: var(--box-shadow-hover);
+}
+
+.testimonial-text {
+  margin-bottom: 2rem;
+  position: relative;
+}
+
+.testimonial-text i {
+  font-size: 3rem;
+  color: var(--primary-color);
+  opacity: 0.2;
+  position: absolute;
+  top: -15px;
+  left: -10px;
+}
+
+.testimonial-text p {
+  color: rgba(255, 255, 255, 0.9);
+  font-style: italic;
+  position: relative;
+  z-index: 1;
+  line-height: 1.8;
+  font-size: 1.1rem;
+}
+
+.testimonial-author {
+  display: flex;
+  align-items: center;
+  gap: 1.5rem;
+}
+
+.testimonial-author img {
+  width: 70px;
+  height: 70px;
+  border-radius: 50%;
+  object-fit: cover;
+  border: 3px solid var(--primary-color);
+}
+
+.testimonial-author h4 {
+  font-size: 1.3rem;
+  margin-bottom: 0.5rem;
+  color: var(--text-color);
+}
+
+.testimonial-author p {
+  color: #aaa;
+  font-size: 1rem;
+}
